Move plugins into their own directory
fix

fix

fix
