+++ packages/react-dom/src/events/plugins/LegacySelectEventPlugin.js
@@ -7,7 +7,7 @@
 
 import {canUseDOM} from 'shared/ExecutionEnvironment';
 import SyntheticEvent from 'legacy-events/SyntheticEvent';
-import isTextInputElement from './isTextInputElement';
+import isTextInputElement from '../isTextInputElement';
 import shallowEqual from 'shared/shallowEqual';
 
 import {
@@ -20,13 +20,13 @@ import {
   TOP_MOUSE_DOWN,
   TOP_MOUSE_UP,
   TOP_SELECTION_CHANGE,
-} from './DOMTopLevelEventTypes';
-import getActiveElement from '../client/getActiveElement';
-import {getNodeFromInstance} from '../client/ReactDOMComponentTree';
-import {hasSelectionCapabilities} from '../client/ReactInputSelection';
-import {DOCUMENT_NODE} from '../shared/HTMLNodeType';
-import {isListeningToAllDependencies} from './DOMEventListenerMap';
-import accumulateTwoPhaseListeners from './accumulateTwoPhaseListeners';
+} from '../DOMTopLevelEventTypes';
+import getActiveElement from '../../client/getActiveElement';
+import {getNodeFromInstance} from '../../client/ReactDOMComponentTree';
+import {hasSelectionCapabilities} from '../../client/ReactInputSelection';
+import {DOCUMENT_NODE} from '../../shared/HTMLNodeType';
+import {isListeningToAllDependencies} from '../DOMEventListenerMap';
+import accumulateTwoPhaseListeners from '../accumulateTwoPhaseListeners';
 
 const skipSelectionChangeEvent =
   canUseDOM && 'documentMode' in document && document.documentMode <= 11;
