+++ packages/react-dom/src/events/plugins/__tests__/BeforeInputEventPlugin-test.js
